// Fitlytics/src/services/generated/models/ExerciseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ExerciseDto
 */
export interface ExerciseDto {
    /**
     * 
     * @type {string}
     * @memberof ExerciseDto
     */
    exerciseId: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseDto
     */
    name: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof ExerciseDto
     */
    muscleGroups: Array<string> | null;
}

/**
 * Check if a given object implements the ExerciseDto interface.
 */
export function instanceOfExerciseDto(value: object): value is ExerciseDto {
    if (!('exerciseId' in value) || value['exerciseId'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('muscleGroups' in value) || value['muscleGroups'] === undefined) return false;
    return true;
}

export function ExerciseDtoFromJSON(json: any): ExerciseDto {
    return ExerciseDtoFromJSONTyped(json, false);
}

export function ExerciseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExerciseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'exerciseId': json['exerciseId'],
        'name': json['name'],
        'muscleGroups': json['muscleGroups'] == null ? null : json['muscleGroups'],
    };
}

export function ExerciseDtoToJSON(json: any): ExerciseDto {
    return ExerciseDtoToJSONTyped(json, false);
}

export function ExerciseDtoToJSONTyped(value?: ExerciseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'exerciseId': value['exerciseId'],
        'name': value['name'],
        'muscleGroups': value['muscleGroups'],
    };
}